requirements.txt
aiodns>=3.0.0
aiohttp>=3.8.4
aiosignal>=1.3.1
astunparse>=1.6.3
async-timeout>=4.0.2
attrs>=22.2.0
beautifulsoup4>=4.11.2
braceexpand>=0.1.7
Brotli>=1.0.9
bs4>=0.0.1
cachetools>=5.3.0
certifi>=2022.12.7
cffi>=1.15.1
chardet>=5.1.0
charset-normalizer>=3.1.0
click>=8.1.3
colorama>=0.4.6
cssselect>=1.2.0
discord-ext-bot>=1.0.1
et-xmlfile>=1.1.0
ffmpeg-python>=0.2.0
Flask>=2.2.3
Flask-SQLAlchemy>=3.0.3
frozenlist>=1.3.3
future>=0.18.3
gitdb>=4.0.10
GitPython>=3.1.31
google>=3.0.0
google-api-core>=2.11.0
google-api-python-client>=2.80.0
google-auth>=2.16.2
google-auth-httplib2>=0.1.0
Google-Images-Search>=1.4.6
googleapis-common-protos>=1.58.0
greenlet>=2.0.2
httplib2>=0.21.0
idna>=3.4
imageio-ffmpeg>=0.4.8
import-expression>=1.1.4
itsdangerous>=2.1.2
Jinja2>=3.1.2
jishaku>=2.5.1
line-profiler>=4.0.3
lxml>=4.9.2
MarkupSafe>=2.1.2
MIDIUtil>=1.2.1
multidict>=6.0.4
numpy>=1.24.2
openpyxl>=3.1.2
orjson>=3.8.7
packaging>=23.0
pandas>=1.5.3
parsel>=1.7.0
Pillow>=9.4.0
playwright>=1.31.1
protobuf>=4.22.1
py-cord>=2.4.0
pyasn1>=0.4.8
pyasn1-modules>=0.2.8
pycares>=4.3.0
pycparser>=2.21
pyee>=9.0.4
pyfiglet>=0.8.post1
PyNaCl>=1.5.0
pyparsing>=3.0.9
python-dateutil>=2.8.2
python-dotenv>=1.0.0
python-resize-image>=1.1.20
pytz>=2022.7.1
pywin32>=305
PyYAML>=6.0
requests>=2.28.2
rsa>=4.9
ruamel.yaml>=0.17.21
six>=1.16.0
slack>=0.0.2
slack-sdk>=3.20.2
slackeventsapi>=3.0.1
smmap>=5.0.0
soupsieve>=2.4
splunk-sdk>=1.7.3
SQLAlchemy>=2.0.5.post1
termcolor>=2.2.0
toml>=0.10.2
typing_extensions>=4.5.0
uritemplate>=4.1.1
urllib3>=1.26.15
w3lib>=2.1.1
Werkzeug>=2.2.3
yarl>=1.8.2
